Otklanjanje gresaka u korisnickom interfejsu, servisima i kontrolerima

+++ Aplikacija za obracun cena gotovih jela/Application/Services/IJeloService.cs
@@ -5,6 +5,6 @@ namespace Application.Services
     public interface IJeloService:IService<JeloDTO>
     {
         Task<decimal> CenaPoPorciji(int id);
-        Task<decimal> UkupnaCena(int id, int brojPorcija);
+        Task<decimal> UkupnaCena(int id, decimal brojPorcija);
     }
 }
